backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "crawl4ai>=0.8.0",
    "fastapi>=0.135.1",
    "litellm>=1.82.2",
    "pdfplumber>=0.11.9",
    "pydantic[email]>=2.12.5",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
    "fpdf2>=2.8.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
